Create confusion matrices when test birds not detected
When one of the test birds is not detected by the model deepforest
encodes the predicted class as -1, which prevents the creation of a
confusion matrix. This change adds a new class for this case with a
value one greater than the current last class, converts the -1 values
to this class, and expands the confusion matrix to include this class.

This allows the confusion matrix to fully show model failures for both
classification and detection.